Implement mrb_context_run.
mrb_context_run is capable of trapping the stack, making it useful in f.e. mirb
and box's repl. A basic implementation and tests are provided.

RunWithContext returns a magic integer that describes the stack in place, so
that it can be re-used on the next call. This is how local variable traverse
ruby parse invocations.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>